+++ src/renderer/main.tsx
@@ -7,10 +7,9 @@ import thunkMiddleware from 'redux-thunk';
 import * as actions from './actions'
 import ApplicationPage from './containers/ApplicationPage'
 import { requireElectron } from './electron';
-import { stateReducer } from './reducers';
 import { State } from './state';
-// import {loadPreferences, updatePreferences} from "./actions";
-
+import { stateReducer } from './reducers';
+import { updatePreferences } from "./actions";
 const electron = requireElectron();
 
 export function main() {
@@ -101,6 +100,14 @@ export function main() {
         event.preventDefault();
         event.stopPropagation();
     });
+
+    window.addEventListener('beforeunload', (event: any) => {
+        event.preventDefault();
+        const state = store.getState();
+        if (state.communication.webAPIClient) {
+            store.dispatch(updatePreferences(state.session) as any);
+        }
+    });
 }
 
 function readDroppedFile(file: File, dispatch: Dispatch<State>) {
@@ -130,4 +137,3 @@ function readDroppedFile(file: File, dispatch: Dispatch<State>) {
         console.warn('Dropped file of unrecognized type: ', file.name);
     }
 }
-